V1.2

+++ src/main/java/Main.java
@@ -8,33 +8,16 @@ import java.sql.*;
  */
 public class Main {
 
-
-    private static ArrayList<Animal> animalList = new ArrayList<>();
-
     public static void main(String[] args) throws java.sql.SQLException, ClassNotFoundException {
 
         Class.forName("org.postgresql.Driver");
-        String jdbcUrl = "jdbc:postgresql://localhost/animals";
+        String jdbcUrl = "jdbc:postgresql://localhost/animals_prod";
         AnimalRepository repository = new AnimalRepository(jdbcUrl);
-        MenuService service = new MenuService(repository);
-
-
-
-
-
         Scanner scanner = new Scanner(System.in);
-
         scanner.useDelimiter("[\n]");
-
-
-        listOfAnimals();
-
         MenuService menuService = new MenuService(scanner);
 
-
-
-
-
+        menuService.displayListOfAnimals(repository.animals());
 
         while (true) {
 
@@ -43,7 +26,7 @@ public class Main {
 
             if (input == menuService.LIST_ANIMALS) {
 
-                menuService.displayListOfAnimals(animalList);
+                menuService.displayListOfAnimals(repository.animals());
 
             } else if (input == menuService.CREATE_ANIMAL) {
 
@@ -59,7 +42,7 @@ public class Main {
 
                 String description = menuService.waitForString("Description of Animal:", true);
 
-                animalList.add(new Animal(name, species, breed, description));
+                repository.createAnimal(new Animal(name, species, breed, description));
 
                 System.out.println("\n" + "Hooray! Your animal has been successfully created.");
 
@@ -67,53 +50,59 @@ public class Main {
 
                 System.out.println("\n-- View an Animal --");
 
-                Animal animal = animalList.get(menuService.waitForInt("Please list the numeric ID of the animal you want to view:")-1);
+                int id = menuService.waitForInt("Please list the numeric ID of the animal you want to view:");
+                ArrayList<Animal> animals = repository.readAnimalByID(id);
 
-                menuService.displayAnimalDetails(animal);
+                if (animals.isEmpty()){
+                    System.out.println("There is no animal with ID: " + id);
+                }else{
+                    menuService.displayAnimalDetails(animals.get(0));
+
+                }
 
 
             } else if (input == menuService.EDIT_ANIMAL) {
 
                 System.out.println("\n-- Edit an Animal --\n");
 
-                int index = menuService.waitForInt("Please enter the numeric ID you would like to edit:")-1;
+                int id = menuService.waitForInt("Please enter the numeric ID you would like to edit:");
 
-                Animal animal = animalList.get(index);
+                ArrayList<Animal> animals = repository.readAnimalByID(id);
 
-                menuService.editAnimal(animal);
+                if (animals.isEmpty()){
+                    System.out.println("There is no animal with ID: " + id);
+                }else{
+                    Animal animal = menuService.editAnimal(animals.get(0));
+                    repository.updateAnimal(animal);
 
-
-
-
-
+                }
 
             } else if (input == menuService.DELETE_ANIMAL)  {
 
-                //NEED TO THROW AN EXCEPTION IF USER CHOOSES AN INVAlID NUMBER
-                //if animalList.size() < index
-                //return waitForInt (prompt: "Please pick an existing ID"
-
                 System.out.println("\n-- Delete an Animal --\n");
 
-                int chooseToDelete = menuService.waitForInt("Please list the numeric ID of the animal that you wish to delete:")-1;
+                int id = menuService.waitForInt("Please list the numeric ID of the animal that you wish to delete:");
 
-                Animal animal = animalList.get(chooseToDelete);
+                ArrayList<Animal> animals = repository.readAnimalByID(id);
 
-                menuService.displayAnimalDetails(animal);
+                if (animals.isEmpty()){
+                    System.out.println("There is no animal with ID: " + id);
+                }else{
+                    menuService.displayAnimalDetails(animals.get(0));
 
-                if(menuService.deleteOrNot("Are you sure that you want to delete this animal? y / n")){
+                    if(menuService.deleteOrNot("Are you sure that you want to delete this animal? y / n")){
 
-                    animalList.remove(chooseToDelete);
+                        repository.deleteAnimal(id);
 
-                    System.out.println("Your animal has been successfully deleted.");
-
+                        System.out.println("Your animal has been successfully deleted.");
+                    }
 
                 }
 
 
             } else if (input == menuService.QUIT_PROGRAM) {
 
-                System.out.println("Exiting program. Your information will not be saved.");
+                System.out.println("Exiting program. Your information will be saved.");
                 break;
 
             } else {
@@ -128,16 +117,16 @@ public class Main {
 
     }
 
-    private static void listOfAnimals(){
-
-        animalList.add(new Animal("Fred", "dog", "Golden retriever", "beautiful"));
-        animalList.add(new Animal("Wilma", "cat", "Persian", "luxurious hair"));
-        animalList.add(new Animal("Juan", "pig", "Cumberland pig", "smells of elderberries"));
-        animalList.add(new Animal("Colleen", "hamster", "dwarf", "slightly cross-eyed"));
-        animalList.add(new Animal("Julius", "dog", "beagle", "one leg is a bit shorter than the others"));
-        animalList.add(new Animal("Vlad", "cat", "Maine coon", "long, gray hair and chubby face"));
-
-    }
+//    private static void listOfAnimals(){
+//
+//        animalList.add(new Animal("Fred", "dog", "Golden retriever", "beautiful"));
+//        animalList.add(new Animal("Wilma", "cat", "Persian", "luxurious hair"));
+//        animalList.add(new Animal("Juan", "pig", "Cumberland pig", "smells of elderberries"));
+//        animalList.add(new Animal("Colleen", "hamster", "dwarf", "slightly cross-eyed"));
+//        animalList.add(new Animal("Julius", "dog", "beagle", "one leg is a bit shorter than the others"));
+//        animalList.add(new Animal("Vlad", "cat", "Maine coon", "long, gray hair and chubby face"));
+//
+//    }
